changes of dashboard and comments are done
